backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
pydantic>=2.10.0
pydantic-core>=2.23.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
cors==1.0.1
